[FIX] #189 회원 정보 수정 후 미반영 문제 해결
